moved firebase/arweave IO code to nonzone-lib
