+++ app/package.json
@@ -30,9 +30,9 @@
         "expo-image-picker": "~8.3.0",
         "expo-linking": "^1.0.1",
         "expo-permissions": "~9.0.1",
-        "expo-splash-screen": "^0.3.1",
+        "expo-splash-screen": "^0.5.0",
         "expo-web-browser": "~8.3.1",
-        "firebase": "7.19.1",
+        "firebase": "7.9.0",
         "nonzone-lib": "1.0.0",
         "react": "16.13.1",
         "react-dom": "16.13.1",
