+++ nonzone-lib/package.json
@@ -24,10 +24,12 @@
         "deploy": "gh-pages -d example/build"
     },
     "peerDependencies": {
-        "react": "^16.0.0"
+        "arweave": "^1.9.1",
+        "firebase": "^7.9.0",
+        "react": "^16.0.0",
+        "smartweave": "^0.3.5"
     },
     "devDependencies": {
-        "microbundle-crl": "^0.13.10",
         "babel-eslint": "^10.0.3",
         "cross-env": "^7.0.2",
         "eslint": "^6.8.0",
@@ -41,6 +43,7 @@
         "eslint-plugin-react": "^7.17.0",
         "eslint-plugin-standard": "^4.0.1",
         "gh-pages": "^2.2.0",
+        "microbundle-crl": "^0.13.10",
         "npm-run-all": "^4.1.5",
         "prettier": "^2.0.4",
         "react": "^16.13.1",
