+++ nonzone-lib/src/firebase/firebase.js
@@ -1,10 +1,10 @@
 import * as firebase from 'firebase/app';
-import 'firebase/analytics';
+// import 'firebase/analytics';
 import 'firebase/auth';
 import 'firebase/database';
 import 'firebase/functions';
-import firebaseConfig from '../../firebaseConfig';
-import firebaseConfigDev from '../../firebaseConfigDev';
+import firebaseConfig from '../firebaseConfig';
+import firebaseConfigDev from '../firebaseConfigDev';
 import { useEffect } from 'react';
 import React from 'react';
 
@@ -14,7 +14,7 @@ const fbConf =
 
 console.log('Init with', fbConf);
 firebase.initializeApp(fbConf);
-firebase.analytics();
+// firebase.analytics();
 
 export const Login = ({ onCancel, onSignedIn }) => {
     const ref = React.useRef();
