+++ nonzone-lib/src/firebaseConfig.js
