+++ nonzone-lib/src/firebaseConfigDev.js
