+++ webapp/package.json
@@ -12,7 +12,7 @@
         "arql-ops": "^1.0.0",
         "arweave": "^1.9.1",
         "axios": "^0.20.0",
-        "firebase": "^7.15.0",
+        "firebase": "7.9.0",
         "firebaseui": "^4.5.1",
         "nanoid": "^3.1.12",
         "nonzone-lib": "1.0.0",
@@ -20,7 +20,7 @@
         "react-dom": "^16.13.1",
         "react-redux": "^7.2.1",
         "react-router-dom": "^5.2.0",
-        "react-scripts": "3.4.1",
+        "react-scripts": "3.4.3",
         "redux": "^4.0.5",
         "smartweave": "https://arweave.net/AM-u4X2po-3Tx7fma3lRonCfLwrjI42IALwDL_YFXBs"
     },
