+++ webapp/src/App.js
@@ -30,7 +30,7 @@ import {
     signOut,
     checkInitialBalance,
     Login,
-} from './api';
+} from 'nonzone-lib';
 import { restoreLastLocation, storeLastLocation } from './utils';
 
 const GOOGLE_API_KEY = process.env.REACT_APP_GOOGLE_API_KEY || '';
